requirements.txt
ultralytics
opencv-python
python-telegram-bot
python-dotenv
numpy
